refactor[#254]: 미사용 SecurityPath 제거 및 레거시 코드 삭제
