--- src/main/java/JJinBBang/app/domain/user/exception/KakaoAuthException.java
@@ -1,25 +1,0 @@
-package JJinBBang.app.domain.user.exception;
-
-import JJinBBang.app.global.error.exception.AuthGroupException;
-
-public class KakaoAuthException extends AuthGroupException {
-    public KakaoAuthException(String message) {
-        super(message);
-    }
-
-    public static KakaoAuthException existAccount(){
-        return new KakaoAuthException("이미 회원가입된 계정입니다.");
-    }
-
-    public static KakaoAuthException accessTokenGenerateError(){
-        return new KakaoAuthException("카카오 API 엑세스 토큰 발급에 실패하였습니다.");
-    }
-
-    public static KakaoAuthException userInfoFetchFailed(){
-        return new KakaoAuthException("카카오 API 엑세스 토큰으로 유저 정보 조회에 실패하였습니다.");
-    }
-
-    public static KakaoAuthException notAllowedRedirectUri(){
-        return new KakaoAuthException("허용되지 않은 리다이렉트 URI 입니다.");
-    }
-}
