--- src/main/java/JJinBBang/app/domain/user/service/OAuthService.java
@@ -1,10 +1,0 @@
-package JJinBBang.app.domain.user.service;
-
-import JJinBBang.app.domain.user.entity.Users;
-
-public interface OAuthService {
-
-    Users login(String oauthProvider, String oauthCode, String redirectUri);
-
-    Users signup(Users user);
-}
